Refactor directory structure and add placeholder for arithmetic game
