+++ src/arithmetic/main.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import HomeButton from '../common/HomeButton';
+
+const Arithmetic = () => (
+    <div id="arithmetic">
+        <HomeButton />
+        <h2>Placeholder for Arithmetic game</h2>
+    </div>
+);
+
+export default Arithmetic;
