+++ src/home/main.js
@@ -1,12 +1,13 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import './Home.css';
+import './styles.css';
 
 const Home = () => (
     <div className="home">
         <header className="home-header">
         Games in React
         <Link className="game-link" to ="/hello">Hello</Link>
+            <Link className="game-link" to ="/arithmetic">Arithmetic</Link>
         </header>
     </div>
 );
